+++ spec/mailers/student_mailer_spec.rb
@@ -1,11 +1,13 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
+include MailHelper
 
 RSpec.describe StudentMailer, type: :mailer do
   # This tests inherited functionality from ApplicationMailer
   let(:user) { build_stubbed(:user) }
   let(:draw) { build_stubbed(:draw) }
+  let(:college) { build_stubbed(:college) }
 
   # rubocop:disable Rspec/ExampleLength
   it 'prepends [Example] to the subject if the user is an admin' do
@@ -25,5 +27,19 @@ RSpec.describe StudentMailer, type: :mailer do
                                                                .intent_deadline)
     expect(message.subject).not_to include('[Example]')
     # rubocop:enable Rspec/ExampleLength
+  end
+
+  context 'vesta links' do
+    it 'includes a Vesta link in intent reminder email' do
+      message = described_class.intent_reminder(user: user, intent_deadline:
+        draw.intent_deadline, college: college)
+      expect(message.html_part.decoded).to include(root_url(host: college.host))
+    end
+
+    it 'includes a Vesta link in group reminder email' do
+      message = described_class.locking_reminder(user: user, locking_deadline:
+        draw.locking_deadline, college: college)
+      expect(message.html_part.decoded).to include(root_url(host: college.host))
+    end
   end
 end