+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 * Fix NoMethodError thrown by AdminDashboard Draw#show ([#965](https://gitlab.com/yale-sdmp/vesta/issues/965)).
 * Remove deprecated User name attribute ([#962](https://gitlab.com/yale-sdmp/vesta/issues/962)).
 * Refactor code to remove unused CSS styles ([#738](https://gitlab.com/yale-sdmp/vesta/issues/738)).
+* Add links to Vesta in the intent and group reminder emails. ([#931](https://gitlab.com/yale-sdmp/vesta/issues/931)).
 
 ## v2.3.3 - 2019-04-01
 ### Fixed